+++ vite-plugin-ssr/node/createPageRenderer.ts
@@ -117,7 +117,9 @@ function assertArguments(
       wrongViteDevServerValueError
     )
     assertUsage(
-      (viteDevServer as any as ViteDevServer).config.plugins.find((plugin) => plugin.name.startsWith('vite-plugin-ssr')),
+      (viteDevServer as any as ViteDevServer).config.plugins.find((plugin) =>
+        plugin.name.startsWith('vite-plugin-ssr')
+      ),
       "`vite-pugin-ssr`'s Vite plugin is not installed. Make sure to add it to your `vite.config.js`."
     )
   }